chat and iconrefmap
